Use wrapper route component to manage lod users import context
