+++ kolibri/plugins/device/assets/src/constants.js
@@ -17,6 +17,7 @@ export const PageNames = {
   MANAGE_SYNC_SCHEDULE: 'MANAGE_SYNC_SCHEDULE',
   EDIT_SYNC_SCHEDULE: 'EDIT_SYNC_SCHEDULE',
   USERS_ROOT: 'USERS_ROOT',
+  USERS_PAGE: 'USERS_PAGE',
   USERS_SELECT_FACILITY_FOR_IMPORT: 'USERS_SELECT_FACILITY_FOR_IMPORT',
   USERS_IMPORT_USER_WITH_CREDENTIALS: 'USERS_IMPORT_USER_WITH_CREDENTIALS',
   USERS_IMPORT_USER_AS_ADMIN: 'USERS_IMPORT_USER_AS_ADMIN',
